requirements.txt
selenium==3.141.0
django-environ==0.4.5
validators==0.18.2
